Fix: Change motor scale to integer
1. We assume the problem of 1 mm error is due to the error of floating
point, need to test

+++ src/Robot/MotorUnion/Motor/motor/motor.h
@@ -35,7 +35,7 @@ protected:
 	/* Conversion of units */
 	float Angle2MotorScale;
 	float MotorScale2Angle;
-	float Rev_Resol_Extended;		// Resolution for one rev in extended mode
+	int Rev_Resol_Extended;			// Resolution for one rev in extended mode
 	float Scale2RPM;
 	float Scale2RPMM;
 
@@ -43,7 +43,7 @@ protected:
 	unsigned char Motor_Operating_Mode;
 	short Motor_CenterScale;
 	float Motor_Angle;			  	// (degree)
-	float Motor_Scale;			  	// Goal Position (Motor Scale)
+	int Motor_Scale;			  	// Goal Position (Motor Scale)
 	int Motor_Velocity;			  	// Goal Velocity
 	int Motor_Accel;			  	// Goal Acceleration
 	float Motor_Present_Angle;	  	// Present Position (degree)
@@ -56,8 +56,8 @@ protected:
 	int Min_Position_Limit;			// Minimum value of position control
 	int Max_Velocity_Limit;			// Maximum value of velocity control
 	int Min_Velocity_Limit;			// Minimum value of velocity control
-	float Max_Extend_Limit;			// Maximum value of extended mode control
-	float Min_Extend_Limit;			// Minimum value of extended mode control
+	int Max_Extend_Limit;			// Maximum value of extended mode control
+	int Min_Extend_Limit;			// Minimum value of extended mode control
 	int Max_Value_In_1_rev;			// Maximum value in one round in extended mode
 	int Min_Value_In_1_rev;			// Minimum value in one round in extended mode
 	int Max_Accel_Limit;
